Big controls update
Rewrote controls internal handling and implemented UI for it + some misc stuff.
Main points:
- now supporting all the controller elements that SDL supports: buttons, hats,
  axes and trackballs; and the UI is there for users to manipulate them
- using SDL game controller mappings to map default controller configuration
- changed the way controls work in menus: now both key sets, mouse and all
  connected joysticks control the menus, and not just selected player 1/2
  controls; and there's an option to disable this
- PC scancodes removed in favor of SDL scancodes
- updated documentation about controls, one file for SWOS internals and the
  other about how the new stuff works
- bumped SDL2 to 2.0.12
- reorganized VC++ projects into subdirectories
- added syntax highlighting for *.mnu files
- unfortunately broke edit tactics menu
- rewrote menu text printing, so it doesn't write those annoying x's in a box
  for unrecognized characters
- also the menu strings will now be automatically elided inside their
  respective item box
- converted most of the menu drawing code, it might be possible to ditch
  VGA-based drawing with some effort in the future
- fixed player keys getting precedence over function keys (stats, etc.)
- fixed options never applying default values if not present in ini file
- fixed mouse clicking in menus: it now triggers when _releasing_ the button
- fixed right mouse button exiting from several main menus, but it's still
  quite menu-dependent so it might fail spectacularly on some menus; consider
  experimental
- plenty of small UI/menu behavior fixes
- repaired test project and brought it to life in 64-bit
- [ida2asm] fixed incorrectly reporting an error about unused symbol when the
    symbol has been eliminated by a range and at the same time marked for
    import from C++
- [ida2asm] fixed invalid VM constant memory write when address % 4 == 3
- [ida2asm] added new directive for exports: (align: <n>)
- [mnu2h] finally fixed tests
- [mnu2h] refactored and fixed a lot of small issues
- [mnu2h] added "endX" entry dot property
- [mnu2h] added "entryAlias" command to define aliases for entries
- [mnu2h] disallowed top-level export
- [mnu2h] added menu entry aliases to C++ Entries enum as well
- [mnu2h] added string indexing and Ruby-like symbols, :abc -> "abc" to
    preprocessor expressions
- [mnu2h] added "int" and "str" conversion operators to p.e.
- [mnu2h] apostrophe ' is now ignored among digits in p.e., just like in C++,
    so it can be used as decimal separator
- [mnu2h] added a warning if p.e. uses empty global/menu variables
- [mnu2h] added missing unary + operator to p.e.
- [mnu2h] added alias "onRestore" for "onReturn" function
- [mnu2h] added #assert preprocessor directive
- [mnu2h] added @nextOrdinal/@nextOrd pseudo-constant to refer to the next
    entry's ordinal
- [mnu2h] added constants for on select mask control values
- [mnu2h] predeclared entry reference variables mirroring Entries enum, like
    headerEntry etc.

- scancodes changed -- you might need to reconfigure keys/delete old ini
- added automatic string eliding in menus, but haven't fully tested behavior in SWOS menu, weird things might
  happen here and there

TODO:
- untested option when all game controllers control menus is disabled (not sure what happens :))
- SDL game controller mappings are largely untested, seems to work fine for now, but will need to test more
  thoroughly sometime
- there is an experimental feature: click right mouse button to exit from a menu, but it's only been tested
  on basic menus, it could fail spectacularly on some other, consider yourself warned
- added some `mnu2h` documentation (for @starwindz), it's still a work in progress but hope it helps

+++ src/audio/SoundSample.h
@@ -22,7 +22,7 @@ public:
     bool hasData() const;
     int chanceModifier() const;
     void setChanceModifier(int chanceModifier);
-    size_t hash() const;
+    unsigned hash() const;
 
     bool operator==(const SoundSample& other) const;
 
@@ -30,9 +30,9 @@ private:
     void assign(const SoundSample& other);
     void convertRawToWav();
     void loadChunk();
-    std::tuple<char *, size_t, bool> loadWithAnyAudioExtension(const char *path, size_t baseNameLength, const char *ext);
+    std::tuple<char *, unsigned, bool> loadWithAnyAudioExtension(const char *path, unsigned baseNameLength, const char *ext);
 
-    static size_t getAudioFileOffset(const char *ext, bool isRaw);
+    static unsigned getAudioFileOffset(const char *ext, bool isRaw);
     static bool isRawExtension(const char *ext);
     static bool is11KhzSample(const char *name);
     static int getMaxAudioExensionLength();
